refactor(tasks): migrate data module Signed-off-by: Cameron Smith <cameron.ray.smith@gmail.com>
