+++ src/pyrovelocity/tasks/__init__.py
@@ -1,0 +1,6 @@
+import pyrovelocity.tasks.data
+
+
+__all__ = [
+    "data",
+]
