+++ src/pyrovelocity/tests/test_data.py
@@ -5,14 +5,14 @@ from pathlib import Path
 import pytest
 import requests_mock
 
-from pyrovelocity.data import _validate_url_and_file
-from pyrovelocity.data import download_dataset
-from pyrovelocity.data import load_anndata_from_path
-from pyrovelocity.data import subset_anndata
+from pyrovelocity.tasks.data import _validate_url_and_file
+from pyrovelocity.tasks.data import download_dataset
+from pyrovelocity.tasks.data import load_anndata_from_path
+from pyrovelocity.tasks.data import subset_anndata
 
 
 def test_load_data_module():
-    from pyrovelocity import data
+    from pyrovelocity.tasks import data
 
     print(data.__file__)
