+++ src/pyrovelocity/train.py
@@ -15,10 +15,10 @@ from mlflow import MlflowClient
 from numpy import ndarray
 from scvi.model._utils import parse_device_args
 
-from pyrovelocity.data import load_anndata_from_path
 from pyrovelocity.io.compressedpickle import CompressedPickle
 from pyrovelocity.logging import configure_logging
 from pyrovelocity.models import PyroVelocity
+from pyrovelocity.tasks.data import load_anndata_from_path
 from pyrovelocity.utils import print_anndata
